Wired up signal connection and device identification
Implemented reactors and the base state to allow for calling join() and
letting a reactive flow continue to connect signal socket, fetch device
credentials, and identify the device on the signal socket.

The state is a bit more verbose than in the PWA, mainly due to the
auto-connect behaviour of the socket in the PWA, meaning in the SDK we
need an extra step (thus extra state) to transition from connected to
identified.